requirements.txt
numpy~=1.23.0
protobuf~=4.21.0
grpcio~=1.51.0
qdrant-client~=1.6.0
